+++ vite-plugin-ssr/node/runtime/renderPage/RenderErrorPage.ts
@@ -3,7 +3,7 @@ export { RenderErrorPage }
 export { assertRenderErrorPageExceptionUsage }
 
 import { assertPageContextProvidedByUser } from '../../../shared/assertPageContextProvidedByUser'
-import { assert, assertUsage, objectAssign, isBrowser } from '../utils'
+import { assert, assertUsage, objectAssign, isBrowser } from '../../utils'
 const stamp = '__isRenderErrorPageException'
 
 assert(!isBrowser())
