+++ vite-plugin-ssr/node/runtime/renderPage/createHttpResponseObject.ts
@@ -20,7 +20,7 @@ import {
 } from '../html/stream'
 import { inferEarlyHintLink } from '../html/injectAssets/inferHtmlTags'
 import type { PageAsset, GetPageAssets } from './getPageAssets'
-import { assert, assertUsage, assertWarning } from '../utils'
+import { assert, assertUsage, assertWarning } from '../../utils'
 import { isErrorPageId } from '../../../shared/route'
 import { getHtmlString, type HtmlRender } from '../html/renderHtml'
