+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets/getManifestEntry.ts
@@ -1,7 +1,7 @@
 export { getManifestEntry }
 
 import type { ViteManifest, ViteManifestEntry } from '../../helpers'
-import { assert, slice, isNpmPackageModule } from '../../utils'
+import { assert, slice, isNpmPackageModule } from '../../../utils'
 import { assertClientEntryId } from './assertClientEntryId'
 
 function getManifestEntry(
