+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets/retrieveAssetsProd.ts
@@ -1,6 +1,6 @@
 export { retrieveAssetsProd }
 
-import { assert, isNpmPackageModule } from '../../utils'
+import { assert, isNpmPackageModule } from '../../../utils'
 import { ViteManifest } from '../../helpers'
 import { getManifestEntry } from './getManifestEntry'
 import { extractAssetsAddQuery } from '../../../plugin/plugins/extractAssetsPlugin/extractAssetsAddQuery'
