+++ vite-plugin-ssr/node/runtime/renderPage/preparePageContextForRelease.ts
@@ -3,7 +3,7 @@ export type { PageContextPublic }
 
 import { isErrorPageId } from '../../../shared/route'
 import type { ExportsAll } from '../../../shared/getPageFiles'
-import { assert, hasProp, isPlainObject, isObject } from '../utils'
+import { assert, hasProp, isPlainObject, isObject } from '../../utils'
 import { sortPageContext } from '../../../shared/sortPageContext'
 import { addIs404ToPageProps } from '../helpers'
 import { assertURLs, PageContextUrls } from '../../../shared/addComputedUrlProps'
